requirements.txt
# 기본 라이브러리
numpy>=1.22.0
pandas>=1.5.0
matplotlib>=3.5.0
seaborn>=0.12.0
scipy>=1.7.0
scikit-learn>=1.0.0
pillow>=9.0.0

# 딥러닝 라이브러리
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.34.0
timm>=0.9.2  # Vision Transformer 관련 모델을 위한 라이브러리

# 이미지 처리 
opencv-python>=4.7.0
albumentations>=1.3.0

# 유틸리티
tqdm>=4.64.0
pyyaml>=6.0.0
tensorboard>=2.12.0
wandb>=0.15.0  # 실험 추적 (선택 사항)

# 데이터 처리
h5py>=3.7.0

# 웹 애플리케이션
flask>=2.2.0

# 시스템 체크를 위한 유틸리티
psutil>=5.9.0  # 시스템 정보 수집
tabulate>=0.9.0  # 테이블 형식 출력